tools: marvell: add printk parser
Add to the rdp tool a printk parser for using in kernel 3.10 and up.

Add the parser tool written by Anton also as a stand alone tool.

The parser is converting the form of variable length records
to a pure text that can be opened with a text editor.

Add of utils.cpp which centralized function
that in use in multiple files.

build_rdp_linux or rdp.dsw(windows)
will generate the 2 executable files.

Change-Id: I2941c887d56c8ba0b22cc6af5da2a9077b1b4cd1
Signed-off-by: Ariel Heller <arielh@marvell.com>