marvell: debug: add mmplog driver
If enable MRVL_LOG, we'll reserve a non-cacheable log buffer
to record all the kernel printk logs.

The buffer would be a ring buffer which keep a position pointer
and magic number "LOGM" at the end of the buffer.

This buffer can be shared with bootloader, so both bootloader
and kernel logs can be recorded in it.

Non-cacheable buffer would be helpful in hang issue debug.

Change-Id: I20873363abe1fd8b83a4eedad310fb758769e0fc
Signed-off-by: Lisa Du <cldu@marvell.com>