drivers: input: add support for delayed key-panic event
This patch expand the ability of keyreset driver to be able to call panic
where some arbitrary key combination is pressed (and not released) for
specified amount of time.
This feature is defined through dts and is backward compatible to act as
before (activate immediately) to platform that didn't add the specific
node in dts.

Change-Id: Iecc401525a601523017ce313c28451740fb390f8
Signed-off-by: Ohad Sharabi <sohad@marvell.com>