clk: mmp: pxa1956: update bit field define for dfc_status
We use bit[7:8] on pxa1936 and bit[7:10]
on pxa1956. As this register is read-only,
just expand the bit field for dfc_cause to
4 bits. On pxa1936, we know the reserved
bits return 0. Actually dfc_cause is not
used in our code.

Also remove enum defines for dfc_cause to
avoid confusion. They are not used as well.

Change-Id: I501f60a4c05917378e4fb276932476f94d1a6d20
Signed-off-by: Bill Zhou <zhoub@marvell.com>