usb: phy: add delay before set PHCD bit.
In helan3, it will cause ~4mA power when boot up without usb pluged in.
then the power diff will not exist after plug in/out usb.
which is because the PHCD bit is not set when usb is not plug in.
we need add delay() before set PHCD bit in order to make axi clock stable,
or the PHCD bit will not be set.

Change-Id: I93b6aefaa8089632d0bc8872d2d2cbdb9deed4b2
Signed-off-by: Wenbing Wang <wangwb@marvell.com>