hwmon: memsic: fix the panic when boot up on debug version of dkb1936
the debug kernel of 1936dkb debug image can't boot up. It is caused by free the
buffer wrong in the driver of mxc400x. From my understanding,on the other
version this buffer is freed and its content just is not changed. So
there is not this issue. At the debug version, the buffer is changed,
this issue occurs.

Change-Id: I472774b2dcb5105e3ff2f7928692ce0abe6120fc
Signed-off-by: Xiaofei Zhu <zhuxf@marvell.com>