cpufreq: mmp: add common files for mmp cpufreq drivers
To support the feature of DDR up threshold, common
files are created to be shared by all platforms.

By default, when cpu frequency is >=800M,
a DDR up threshold 30 will be sent to
MCK driver.

User can modify the parameters by debugfs:

For dual clusters,
/sys/kernel/ddr_upthreshold/upthrd_clstx
/sys/kernel/ddr_upthreshold/upthrd_swap_clstx
x is cluster index, 0 or 1.

For single cluster,
/sys/kernel/ddr_upthreshold/upthrd_cpu
/sys/kernel/ddr_upthreshold/upthrd_swap_cpu

The valid range for DDR up threshold is (0, 100).
The unit for cpu swap frequency is KHz.

Change-Id: Ic6ea4e3b940abda50417ab97d30d57301c5675bd
Signed-off-by: Bill Zhou <zhoub@marvell.com>