mm: page_alloc: fix the CMA migratetype bug.
We always receive the “alloc_contig_range test_pages_isolated
failed” warning reports.

It is because in the __rmqueue_highest_cma() function, a single
page allocated don't set the corresponding migratetype according as its pageblock migratetype.

For the single page(order=0), the kernel store and check its migratetype via its page->index.
Please refer the free_pcppages_bulk() function.

The page->index default value is 0. Unfortunately, the MIGRATE_UNMOVABLE value is 0.
So the page allocated from CMA area via __rmqueue_highest_cma() function will be taken as
the unmovable page and in the end is freed into the unmovable free list.

In the alloc_contig_range()  function, these unmovable pages will be skipped , not to be migrated
even they are still referenced.
This situation will result in the failure of test_pages_isolated() function and producing the
“alloc_contig_range test_pages_isolated failed” warning.

Change-Id: Ie1752b37359fd17041233ca17b8d138f5faa5229
Signed-off-by: Zhigang Hu <gavinhu@marvell.com>