power: 88pm88x: fix system can not suspend error
after system resumed by console input, system will stay awake
and wait for a long time.
it's because we will keep system alive by holding wakeup event
until we release it in the delayed work - monitor work which
is used to update battery status. but it's a deferrable work
which maybe delayed for a long time.
in order to handle this error, we use non-deferrable work instead
of deferrable one.

Change-Id: Ib2e0d15d81577e76f334a6411cac0a8a666080fd
Signed-off-by: Fenghang Yin <yinfh@marvell.com>