drivers: char: perform auto-calibration in background
Vibrator calibration process is triggered by SW and takes the HW few
hundreds mili-seconds to complete.
Currently the code polls the HW in probe context in order to check if
calibration completed and so increases kernel's boot time significantly.
This patch performs the calibration in background by requesting
deferred probe when it triggers the calibration and by that allows the HW
to complete the calibration in background while other devices are probed.

Change-Id: I0b11289d329acc32d3549b2b208ad23a2c3e5004
Signed-off-by: Ohad Sharabi <sohad@marvell.com>