mfd: 88pm88x: remove battery detection interrupt
since the battery detection interrupt is not used anymore, we need
remove it from the platform resources.
otherwise the info->irq_nums in 88pm88x_battery.c is 3, but the array
size of pm88x_irq_descs[] is 2, there is potential error to access
the address of pm88x_irq_descs[2], because it does not exist at all.

Change-Id: Iae31eb9a9bf09a69ed5cf9bc1ab271e70479238d
Signed-off-by: Zhongbo Han <zhbhan@marvell.com>