mmc: host: sdhci: fine tune timeout_timer for IRQ pending
In the real case, we often meet a problem:
Even SDH has asserted its IRQ, as for some reason, CPU's IRQ has been
disabled for a long time or sdhci_tasklet_finish is not scheduled. As a
result, the sdhci_timeout_timer is called.

In this case, problem happens in cpu/system side but not SDH side; and
SDH is victim in fact.

This patch, we add some protection for this case and report this
situation. Also let SDH to wait CPU/System to handle IRQ.

Change-Id: Ib93cc684e614d409105071178b0ffb6176a34cfb
Signed-off-by: Jialing Fu <jlfu@marvell.com>