drivers: emmd: add interface to enable/disable the emmd function
the emmd function is enabled by default, so when a kernel panic is
encountered, system will reboot and enter emmd mode.
however we may not want our end user to see it. it's better that system
should just reboot if there is a kernel panic when the phone is deployed
to end user.

an interface is added to enable/disable the emmd function, the usage is:
enable emmd dump:
    sysctl -w kernel.emmd_enable=1
    or
    echo 1 > /proc/sys/kernel/emmd_enable
disable emmd dump:
    sysctl -w kernel.emmd_enable=0
    or
    echo 0 > /proc/sys/kernel/emmd_enable

Change-Id: Ia6ffc15e408b42a475aa1f9fe476a74f22720228
Signed-off-by: Jinfeng Chen <jfchen@marvell.com>
Signed-off-by: Zhongbo Han <zhbhan@marvell.com>