drivers: gpio: Request mmp GPIO interrupt to GIC controller
GPIO interrupt is not only a secondary interrupt controller but also
a sub interrupt of GIC. Current code only uses irq_set_chained_handle
to enable it without calling irq_set_affinity. Consequently, the GPIO
interrupt will use the default cpu routing of GIC(and ICU maybe). If
the defalut affinity value of GIC is zero, then this interrupt can't
be routed to any cpu, which is not expected.

This patch uses request_irq to register gpio interrupt to GIC. This
function will defaultly call irq_set_affinity so that GPIO interrupt
can be routed to a cpu available.

Change-Id: I682b6155a6eb45316313441df2a6f9629529a2e4
Signed-off-by: Fangsuo Wu <fswu@marvell.com>