power: 88pm88x_battery: support series resistor connected to BAT_ID
in order to protect from ESD, a series resistor may be connected to BAT_ID
together with a capacitance. when we use ntc-table, the series resistor
should be added.

a new battery parameter named "ntc-series-ohm" is added for this purpose.
if the series resistor is not used, we can set it to 0 or delete it.
otherwise set it to the right value.

Change-Id: I67499c45903443d4e5ac000567ff7a200d81a00d
Signed-off-by: Zhongbo Han <zhbhan@marvell.com>