power: 88pm88x_charger: print the real charger type
the type of power supply is only used by user space. when the healthd is not
initialized, we have to map the real charger type to the type of power supply.
so the type of power supply can't indicates the truth.

the real charger type should be used to debug.

Change-Id: Ia2cca6856df2f002743c4170cfbee25756006ac5
Signed-off-by: Zhongbo Han <zhbhan@marvell.com>