arm64: dts: pxa1936: add quick to set cpuidle latency
On some chips, such as: pxa1936
The C2 latency value may impact eMMC RX timing tuning result.

If set the C2 latency as a big value, the timing tuning restult
became unstable. So we need to keep the C2 latency as small value
during tuning stage and recover the normal value after it.

We add a quick for such chip to call the latency set and recovery
function.

Change-Id: I186baaea17e2c3c148403775ba9b8d4dabba1343
Signed-off-by: Tim Wang <wangtt@marvell.com>