arm64: dts: improve charging temperature
the temperature range 0-40C is suggested to charge according to the battery
spec. and the battery spec also describes that the battery can be charged in
0-50C.

under some stress test, the battery temperature arises and the charging will
be stopped when the temperature is greater than the high temperature threshold.
so we try to improve this threshold.

Change-Id: Icce00d664d27fd4e308a6328e95f46fded1a56c4
Signed-off-by: Zhongbo Han <zhbhan@marvell.com>