arm64: dts: limit the booster current and voltage of flash/torch
the high voltage and large current of flash/torch booster may damage the PMIC.
so 4500mV/1000mA is suggested on 88PM886/88PM880.

but the booster current also depends on the LED spec. so we suggest that
booster current should be in the safe range according to the LED spec.

what's more, if CF_OUTA and CF_OUTB are connected together on some platform,
we suggest to enable "conn-cfout-ab" feature.

Change-Id: Ibde6e97ca13f98acd87bb069830faf7b50be2b5f
Signed-off-by: Zhongbo Han <zhbhan@marvell.com>