input: touchscreen: gt9xx: disable the gesture feature
disable the gesture feature. and disable the ldo when the touchscreen
is suspended. But in this driver, the ldo enable function is called sevaral times.
But disable function is called less than enable function. So in our ldo enable,
disable mechanism, the enable function and the disable function shall be one by one.
So this driver can't disable this ldo. For this reason,a new mechanism is added
in this driver to resolve this problem.

Change-Id: Id0f6b22b9aa1667f18b854406298ea06a964b1c8
Signed-off-by: Xiaofei Zhu <zhuxf@marvell.com>
(cherry picked from commit a7a0b8e390c7b8132e347e59eb3c0e6f7a609d14)