arch: arm: Fix bug D2 can't be entered if ever holds D2 QOS in suspend
The bug that D2 can't be entered is found following below sequence:
1. System enters suspend while holding a D2 QOS, like FM playback
or calling. Since the D2 QOS won't be checked in suspend, suspend
can be triggerd but SOC can only enter D1. However, after the suspend,
*real_idx will change to D1, that means mmp_suspend->suspend_state
is also D1, but it should be always D2;
2. When D2 QOS is released, system re-enters suspend and gets the
LPM status, this time the status is D1. So system can't enter D2.

Change-Id: Ib584672fb54794b626b13bc2343e8b3e65c66bf8
Signed-off-by: Fangsuo Wu <fswu@marvell.com>